moved all user interface functionality to a new Interface class
Modularized the user interface. All UI functionality is moved into a new Interface class. At startup, the program creates an instance of Interface. All input and output to the user is conducted by calling functions on that instance. The UI of the base Interface class is text-only, but future subclasses of Interface could provide graphical interfaces.